Move library to root of repository
GitHub automatically generates .zip files for every commit and release. In order to use these .zip files with the Arduino IDE's Sketch > Include Library > Add .ZIP Library feature it is required for the library to be in the root of the repository, otherwise the process fails "Specified folder/zip file does not contain a valid library".

Inclusion in the Arduino Library Manager index requires the library to be located in the root of the repository.

For these reasons, storing the library in the root of the repository has become the standard for Arduino libraries. Arduino users are most likely to be comfortable working with repositories with this structure.